Moved 'rpm_filter' into 'flight'.
