+++ src/main/osd/osd_elements.c
@@ -81,6 +81,7 @@
 #include "flight/imu.h"
 #include "flight/mixer.h"
 #include "flight/pid.h"
+#include "flight/rpm_filter.h"
 
 #include "io/beeper.h"
 #include "io/gps.h"
@@ -100,7 +101,6 @@
 #include "sensors/battery.h"
 #include "sensors/esc_sensor.h"
 #include "sensors/sensors.h"
-#include "sensors/rpm_filter.h"
 
 
 #define AH_SYMBOL_COUNT 9
