+++ src/main/sensors/gyro.c
@@ -66,6 +66,11 @@
 #include "fc/config.h"
 #include "fc/runtime_config.h"
 
+#ifdef USE_GYRO_DATA_ANALYSE
+#include "flight/gyroanalyse.h"
+#endif
+#include "flight/rpm_filter.h"
+
 #include "io/beeper.h"
 #include "io/statusindicator.h"
 
@@ -73,10 +78,6 @@
 
 #include "sensors/boardalignment.h"
 #include "sensors/gyro.h"
-#ifdef USE_GYRO_DATA_ANALYSE
-#include "sensors/gyroanalyse.h"
-#endif
-#include "sensors/rpm_filter.h"
 #include "sensors/sensors.h"
 
 #if ((FLASH_SIZE > 128) && (defined(USE_GYRO_SPI_ICM20601) || defined(USE_GYRO_SPI_ICM20689) || defined(USE_GYRO_SPI_MPU6500)))
